Move Audit stuff into it's own classes, etc...
