+++ app/Audit/Audit.php
@@ -1,0 +1,8 @@
+<?php
+namespace App\Audit;
+error_reporting(E_ALL ^ E_DEPRECATED);
+
+interface Audit
+{
+    public function audit_results($packages, $vulnerabilities, $output);
+}
